Moved Particle to owen file
